requirements.txt
Pillow
fontawesome-free